user

+++ nodo.h
@@ -10,7 +10,7 @@ void transaction_print (transaction d){
 
 }
 
-transaction* new_transaction(int timestamp, int sender, int receiver, int amount, int reward){
+transaction* new_transaction(long timestamp, int sender, int receiver, int amount, int reward){
     transaction *d = malloc(sizeof(transaction));
     d->timestamp = timestamp;
     d->sender = sender;
